Remove EditPage.css Remove handleSetState method Move validateEmail and validatePassword to util.js Fix minor errors
