+++ src/sass/EditAccount.scss
@@ -10,11 +10,6 @@ h2 {
     display: inline-block;
     vertical-align: top;
 
-    // h2 {
-    //     // color: $pallete-blue;
-    //     margin-bottom: 0.5em;
-    // }
-
     ul {
         list-style: none;
         padding: 0px;
